Refactor MyProfile
